Only support xbar on darwin
